requirements.txt
anthropic>=0.40.0
chainlit>=1.3.0
mcp>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
requests>=2.31.0
